feat: 와인 상세 조회 api 구현 완료

+++ backend/src/main/java/com/ssafy/wible/repository/WineRepository.java
@@ -4,5 +4,5 @@ import com.ssafy.wible.model.entity.Wine;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface WineRepository extends JpaRepository<Wine, Long>, JpaSpecificationExecutor<Wine> {
+public interface WineRepository extends JpaRepository<Wine, Integer>, JpaSpecificationExecutor<Wine> {
 }
